webapp/requirements.txt
flask==3.1.0
gunicorn==23.0.0
paho-mqtt==2.1.0
